CL Fixes
